Implement big query write operation via lambda
