+++ lambda/requirements.txt
@@ -1,0 +1,2 @@
+protobuf==4.21.12
+google-cloud-bigquery==3.4.1
